serveur_web/requirements.txt
flask==0.12.2
flask-restful==0.3.6
blinker	
gevent	
gunicorn	
psycopg2
flask_cors
google-api-python-client==1.6.4
google-auth==1.2.0
google-auth-httplib2==0.0.2
